Fixed support for synchronous calls to update()
If any render handler has a length of 1, it will be provided with a function
that when called, will schedule an update and/or rerender. The expectation is
that this function would be called sometime in the future (usually on the next
anmiation frame).

This operation has generally been working well, but only if calls to this
update function happen *after* the render operation is complete.

Here's a simple example of this process working well:

```javascript
dom.div((update) => {
  doSomethingAsynchronous()
    .then(() => {
      // Mutate models or state
      update();
    });
});
```

Here's an example of this not working well:

```javascript
let firstRun = 0;
dom.div((update) => {
  if (firstRun === 0) {
    firstRun++;
    dom.div({id: 'first'});
    update();
  }

  dom.div({id: 'second'});
});
```

In this second case, the update operation was being discarded and any future
update operations would lose functional scope to the originally-rendered
elements. This appeared as if event handlers were no longer firing on user
interactions and as some features were not rendering, while other features
might.

This is fixed, in that synchronous calls to update within a render pass will
be scheduled and rendered in the future.

This fix simply involved checking the nomplate element hasUpdateableHandler
field (which is availably immediately) instead of the DOM element.render field
which we apply to the element *after* the handler has been instantiated and
configured. The builder was originally checking for this *inside* the function
that creates it.

NOTE: There be some dragons in this code path, this is unlikely the last time
we'll see issues in here.