Introduced support for `dom.external(domElement)`
This is intended to support raw DOM Elements that are long-lived and
often used by external libraries (e.g., three.js).

This is basically a trapdoor that allows an application author to
instantiate and control a nested element tree within a Nomplate view.

To use this, provide the element to the view (usually in options hash)
and then send it as the argument to dom.external(elem) as follows:

```javascript
const dom = require('nomplate').dom;

function myView(options) {
  return dom.div({id: 'my-view'}, () => {
    dom.h1('Hello');
    dom.external(options.domElement);
  });
}
```

When this gets rendered (as elements), you'll see the external element
in the view without async flashing.

NOTE: Also fixed compile warning for NaN comparison that was always
false.